+++ test/support/controllers.ex
@@ -4,21 +4,21 @@ defmodule AegisTest.PuppyController do
 
   use Aegis.Controller, except: [:index]
 
-  action_fallback AegisTest.FallbackController
+  action_fallback(AegisTest.FallbackController)
 
   def index(conn, _params) do
-    text conn, "showing pup(s)"
+    text(conn, "showing pup(s)")
   end
 
   def show(conn, %{"id" => 3}, _user) do
-    text conn, "showing pup"
+    text(conn, "showing pup")
   end
 
   def show(conn, %{"id" => id}, user) do
     puppy = %Puppy{user_id: id}
 
-    with {:ok, conn} <- authorized?(conn, user, puppy, :show) do
-      text conn, "showing pup"
+    with {:ok, conn} <- authorized?(conn, user, puppy, action: :show) do
+      text(conn, "showing pup")
     end
   end
 
@@ -33,11 +33,11 @@ defmodule AegisTest.PuppyWithoutExcludedActionsController do
 
   use Aegis.Controller
 
-  action_fallback AegisTest.FallbackController
+  action_fallback(AegisTest.FallbackController)
 
   def index(conn, _params, user) do
-    with {:ok, conn} <- authorized?(conn, user, Puppy, :index) do
-      text conn, "showing pup(s)"
+    with {:ok, conn} <- authorized?(conn, user, Puppy, action: :index) do
+      text(conn, "showing pup(s)")
     end
   end
 
@@ -49,7 +49,7 @@ end
 defmodule AegisTest.FallbackController do
   use Phoenix.Controller, namespace: AegisTest
 
-  def call(conn, {:error, :not_authorized}) do
-    text conn, "not authorized"
+  def call(_conn, {:error, conn}) do
+    text(conn, "not authorized")
   end
 end